Improve usage of struct member level
Before .level == .mask could mean either full or almost full, depending on the members .write and .read. This design decision was supposed to reduce ram usage as much as possible, but was responsible for more if-statements in the code. Now .level can be .mask+1, but this requires a larger datatype for fifos with a depth 256 or larger, which is chosen automatically. Overall, the new code reduces the flash usage of the test code by 25%!